requirements.txt
numpy~=1.26.4
opencv-python
scikit-image~=0.25.2
Pillow~=11.0.0
scipy~=1.15.2
matplotlib~=3.10.1
torch~=2.5.1+cu121
torchvision~=0.20.1+cu121

pandas~=2.2.3
tqdm~=4.67.1
scikit-learn~=1.6.1
setuptools~=65.5.0
Pygments~=2.19.1
pip~=25.0.1
wheel~=0.45.1
protobuf~=4.25.6
Jinja2~=3.1.4
filelock~=3.13.1
pytz~=2025.2
PyWavelets~=1.8.0
pytest~=8.3.5
requests~=2.32.3
rich~=14.0.0
keras~=2.15.0
colorama~=0.4.6
charset-normalizer~=3.4.1
typing_extensions~=4.12.2
threadpoolctl~=3.6.0
mpmath~=1.3.0
sympy~=1.13.1
astunparse~=1.6.3
tensorboard~=2.15.2
optree~=0.15.0
tensorflow~=2.15.0
packaging~=24.2
opt_einsum~=3.4.0
urllib3~=2.4.0
fsspec~=2024.6.1
MarkupSafe~=2.1.5
joblib~=1.4.2
logger~=1.4
python-dateutil~=2.9.0.post0
imageio~=2.37.0
tifffile~=2025.3.30
cycler~=0.12.1
lazy_loader~=0.4
networkx~=3.3
six~=1.17.0
contourpy~=1.3.2
fonttools~=4.57.0
pyparsing~=3.2.3
kiwisolver~=1.4.8
ImageHash~=4.3.2